requirements.txt
chromadb
sentence-transformers
PyPDF2
google-generativeai
rich
pyyaml
requests
